Convert e2e test script logic into python code
also refactor some of the e2e helper code into utils modules
